refactor(tests): reorganize test utilities into dedicated directory
